+++ test/cases/collection_clone.py
@@ -9,9 +9,11 @@ class CollectionCloneTestCase(TestCase):
 
     def run(self):
 
-        # collection|c clone|c <url> [collection-name] - Clone a collection from a git repository
+        # collection|c clone|c <url> [collection-name] [flags] - Clone a collection from a git repository
         #   url - The URL of the git repository to clone
         #   collection-name - The name of the collection to create. Default is the name of the repository
+        #      Flags:
+        #        -print-path|-P - Print the path of the cloned collection
 
         MSH.run_fail("collection clone")
 
@@ -26,3 +28,9 @@ class CollectionCloneTestCase(TestCase):
         )
         MSH.HOME.add_dir("test-coll").add_file("msh-test-script").check()
         MSH.HOME.add_dir("test-coll").add_file("msh-test-script2").check()
+
+        path =  MSH.HOME.add_dir("test-coll2").get_abs_path()
+        
+        MSH.run_suc(
+            "collection clone https://github.com/cophilot/msh-test-collection.git test-coll2 -print-path"
+        ).has_line(path)